Move components to index.js
